initialize project
the core idea behind this project is to generate a blog post from a GitHub repository's commit history.
I have a lot of repositories and my biggest problem is remembering what I've done in each one of them, especially if I go back to
write the blogpost after a long time. I literally forget what I went through and it's pretty tough to draw a nice story through it.

But also commits are the best way to tell a story about a project, because they are the most detailed and accurate way to tell a story about a project.
it goes through the evolution etc