add base endpoint to scrape commits from repo
this adds a base endpoitn and the logic to get the commits from a repo.

A big problem in this is that the endpoint ion github is rate limited if youre not authenticated.
And we dont have enough time to make the oauth logic. So it's just going to use my personal token.

topic: