improve the personal prompt
this is really important to give context on what your existing blogposts are. this can drive the
generation to sound like you. And it can use Comet to do the fine-tuning of the model to your writing style. It can
eval and help you imrpvoe the style so that you have a more consistent writing style.